Cool, you can now bundle up a page or file in a module as a Base64 encoding, and include it in a menu section of a module's XML definition.  This means that we can have a small base system where even features that are a part of it normally can be added and removed.
I've also added a variable in $_SESSION['revision'] which is the SVN revision for the current checkout (mainly useful if SmoothOperator has been checked out to web root).  Using this we can check for available updates and add items to the database or make modifications if someone is using an older version.