Reasonably large commit - the module installation now supports base64 encoded .tar.gz archives of packages to install, decompresses them, adds menu items, runs a patch (using a unified diff encoded with base64) to change any files to use the SmoothOperator config system and installs the files.  It also saves the file names to the DB so that they can be removed on uninstallation.  Config options are not currently removed.
So, for example, with the Areski CDR Stats package it does this:

1. Download XML File
2. Parse it for archives
3. Parse it for patches
4. Parse it for menus
5. Parse it for config options

Next it installs the above.

Menus appear as menus in the system (and can be moved around etc like normal menus), config options get added to the sourcecode via the patch - i.e. in CDR stats there is a file lib/defines.php where you need to specify DB connection details hard coded.  It patches this so that the config options come from $config_values (stored in $_SESSION), and the options appear in the config section of the main site.  So for example it provides $config_values['cdr_host'].

archive.php allows us to deal with .tar.gz files and extract them without using the base operating system

phppatcher.php allows us to patch the directory structure of the installed files.

At the moment, the cdr directory comes in SVN and this is the subdirectory that is used to store the files - maybe it would be better to use a modules/ directory.